requirements.txt
django
djangorestframework
pillow
djangorestframework-simplejwt
asgiref
coverage
django-cors-headers
pytz
sqlparse
django-environ
django_heroku
gunicorn
drf-yasg
psycopg2
django-environ
django-storages
boto3
sendgrid
twilio